#25 extract validation functions and image upload functions
